Separate DOM logic from the rest, add project view
Move all DOM related stuff to DOMManipulator (creating DOM-elements,
handle DOM events like 'click' on buttons)

There is only one entry point (index.js) and only one output (main.js)
because `pubsub` needs to be a global object. The reason for doing this
is that I subscribe a handler to an event in index.js and publish event
in DOMManipulator.js. So, it should be managed by the same pubsub.

Perhaps I misunderstood something, and there is a way to accomplish it
somehow else.